refactor(pro): extract shared Featherwork Pro paywall from DrillHubPanel
One subscription (RevenueCat 'pro' entitlement) now presents one paywall,
reusable outside the drill hub. Copy is retitled Featherwork Pro to match
the app rename and to cover the customization styles it also unlocks.
